inference-server/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
opencv-python
numpy
Pillow
requests
Cython
tensorflow
flask_cors

# numpy==1.16.0rc1
fastai
starlette
uvicorn
python-multipart
aiofiles
aiohttp